修复在taskqueue为空的时候获取任务导致系统崩溃的问题

+++ TaskQueue.py
@@ -6,11 +6,13 @@ db = server_db.db
 '''
 任务队列实际上应该用redis等nosql数据库来做，这里稍微简化一下...
 '''
+
+
 class TaskQueue(db.Model):
     __tablename__ = 'task_queue'
     id = db.Column(db.Integer, primary_key=True)        # 任务id
     download_url = db.Column(db.Text)                   # 任务资源下载地址
-    target_node = db.Column(db.Integer,default=0)       # 目标节点
+    target_node = db.Column(db.Integer, default=0)       # 目标节点
     upload_url = db.Column(db.Text)                     # 结果上传地址
 
     def set_download_url(self, download_url):
@@ -22,6 +24,12 @@ class TaskQueue(db.Model):
     def set_target_node(self, target_node):
         self.target_node = target_node
 
+    def new_empty_task(self):
+        self.id = None
+        self.download_url = None
+        self.target_node = None
+        self.upload_url = None
+
     def add(self):
         db.session.add(self)
         db.session.commit()
@@ -30,10 +38,15 @@ class TaskQueue(db.Model):
         db.session.delete(self)
         db.session.commit()
 
-    # 获取队列中的第一项，然后在列表中删除这一项
+    # 获取队列中的第一项，然后在队列中删除这一项
+    # 如果队列为空，则返回None
     def get(self):
-        self = self.query.first()
-        self.delete()
+        _task = self.query.first()
+        if _task is not None:
+            self = _task
+            self.delete()
+        else:
+            self.new_empty_task()
         return self
 
     # 转换成字典供给flaskrequest转换成json
@@ -52,5 +65,3 @@ class TaskQueue(db.Model):
             self = _task
             self.delete()
         print('清理数据库成功...')
-
-
